requirements.txt
matplotlib==3.1.1
pandas==0.25.0
biopython==1.74
numpy==1.17.0
scipy==1.3.0
panel==0.6.2